Clean (mostly...) re-implementation of hackaton archievements: implemented generic solution for getUserProperty and added setUserProperty by mapping them onto regular named properties. Made separate classes out of experimental UPnP device implementation.
Note: live adding of new devices did work at the hackaton only due to a weird chain of accidents and oversights. For now, live detection (after collecting at daemon start) is not yet re-implemented, but a // TODO: in upnpdevicecontainer.cpp describes the plan...

Note: the UPnP device specific property "descriptionURL" is now available as a regular device property, and as a workaround also via getUserProperty(0).