protobuf processing: found really nasty bug causing some protobuf messages to be processed twice, and others ignored on vdc side.
Problem was that protobuf-c's xxxx__message__unpack, given a buffer containing two messages, does not error out and also does not (as one might assume) process the first message, but apparently *skips* the first message and then returns the unpacked version of the second.

This happened not often, but apparently pretty reproducably when vdsm sent getProperty for "sensorDescriptions" (both on rascow's Linux and my OS X), processing the getProperty twice. The same bug also caused vdcd to ignore callMinScene on the digiEsp platform.