vDC API: removed session timeout.
The vDC session timeout is not compatible with the vdsm's alive check
which also counts pushProperty messages from the vdc as a life sign,
and thus does not ping devices that have recently pushed a value.

This could lead to session timeouts with vdcs hosting only devices with
a sensor - these devices always send pushProperty and thus
prevented the vdsm to send pings which would have kept the connection
open.

The timeout was specified in the early vDC API drafts, but was not even mentioned in the current vDC specs any more.