Channel sync: usually only store synced value if no new value is already pending to get applied.
This is important for devices like hue which must confirm apply early (when sending out the request), but get a result from hardware with new values later.

Therefore, the channel values might change again before the result arrives. In this case, actual hardware values are already outdated when they arrive and must NOT be stored.