DALI: added more plausibility checks for Bank 0 data, now including GTIN check digit validation and too-many-0xFF patterns.
Reason is that AGAIN a product (linealight.com/i-LÈD/eral LED-FGI332) has fooled the existing checks. This particular product has garbage GTIN, and we hoped to catch it with GTIN check digit validation. Unfortunately, this device's garbage GTIN happens to have the correct check digit. But serial is 0x0000FFFF and GTIN is 0x1FF02FFFFFF, so we catch that with a new suspicuous-total-of-0xFFs-plus-3-or-more-consecutive-equal-bytes heuristic ;-)

Side note: some research has shown that sample libraries for DALI clients from Atmel and Microchip BOTH implement Bank0 memory, but DO NOT have a mechanism to provision GTIN+Serial at production time (Bank 0 contents are just #defines in a .c file!!). This might be the reason why so many DALI device have garbage Bank0 content...