External Devices: Added new device class, which allows external devices living in a separate process/app/script and connecting via a local socket or TCP port.
The external device API allows the connecting device process to define the specifics
of the device as a JSON message, and then communicating device events also in JSON
or optionally in a extremely simple text format (to allow implementation with a few
lines of C without any JSON parser required).

This mechanism allows implementing full featured custom devices in any language
which allows opening a TCP or local socket connection.