Prepared to allow migrating config data from an original unit to a replacement unit without derived dSUIDs changing.
vdc host dSUID generated from MAC address is now persisted in DB. This allows migrating the DB to another unit, while retaining the original vdc host dSUID and all dSUIDs that are derived from it.

A warning is logged when the stored dSUID is different from the MAC generated dSUID, to signal that the vdcd operates in a "device replacement" mode. This should only happen if the replaced device is not in use any more (because the new one now uses the same dSUIDs the old one did - instead of its own, unique, MAC-derived dSUIDs).