Re-added generic vdc sources as p44vdc, adapted XCode project and Makefile.am accordingly.
Note - the name was changed from "libp44vdc" to just "p44vdc" because the
p44 vdc framework is not built as a library, but usually just included
at the source level (as in vdcd).

It is still possible to build a "libp44vdc" based on "p44vdc" sources when
this is actually needed for a specific target platform, but that would
happen in a separate project/repository.

In general, including p44vdc as sources into a main project like vdcd is
easier to work with and maintain, and more efficient at runtime.

But there might be special cases, such as a need for having multiple
vdchosts running on the same platform for modularity reasons - then it
might make sense to share the large part of common code in a shlib.