xcode project: Transition from mongoose to civetweb
httpcomm unit tests now available for http, https with cert and hostname checking, as well as http auth.

We are using programmatic SSL lib loading now, which means that no libssl or libcrypto need to be linked with the executable at build time.

But with NO_SSL_DL=1, vdcd and tests can also be built linking the libs normally.

Note: Use with OpenSSL 1.1 is not yet tested (but prepared in sources)