Added "--allowcloud" command line option for optionally allowing use of cloud services not explicitly configured or obviously expected by the user, such as N-UPnP.
Basic idea is that vdcd/p44vdc should not make any non-local network connections the user has not explicitly configured, like the use of N-UPnP when the user expects hue bridge connection to be a strictly local thing.

On the other hand, vdcs that are designed to integrate obviously cloud based devices (e.g. Netatmo) do not need --allowcloud.