p44script event handling improved with "frozen" values, local functions, signals with data, isok(), ifok(), strrep()
The "frozen" event value mechanism ensures that data objects sent along with an event is returned in the trigger expression evaluation where the event sourcing function/member is used.

Before, the trigger expression evaluation accessed the event sourcing function a second time, expecting it to return the same value as delivered with the event. This required ugly intermediate storage of "last accessed" values, which could cause inconsistencies when triggers fired too rapidly in succession. This is now solved